refactored functions
