+++ package.json
@@ -24,5 +24,6 @@
   "dependencies": {
     "discord.js": "^11.1.0",
     "googleapis": "^19.0.0"
-  }
+  },
+  "devDependencies": {}
 }
